requirements.txt
asgiref==3.9.1
bcrypt==4.3.0
blessed==1.21.0
botocore==1.39.17
cement==2.10.14
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.5
decorator==5.2.1
Deprecated==1.2.18
dj-database-url==3.0.1
Django==5.2.6
django-admin-honeypot-updated-2021==1.2.0
django-admin-thumbnails==0.2.9
django-ipware==7.0.1
django-jazzmin==3.0.1
django-js-asset==3.1.2
fabric==3.2.2
gunicorn==23.0.0
idna==3.10
invoke==2.2.0
jmespath==1.0.1
packaging==24.2
paramiko==3.5.1
pathspec==0.12.1
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.22
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-ipware==3.0.0
PyYAML==6.0.2
requests==2.32.4
semantic-version==2.10.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
termcolor==2.5.0
urllib3==1.26.20
wcwidth==0.2.13
whitenoise==6.11.0
wrapt==1.17.2
